Patient Edit Component Enhancment

+++ src/app/profile/employee-profile/employee-profile-card/employee-profile-card.component.ts
@@ -44,8 +44,10 @@ export class EmployeeProfileCardComponent implements OnInit {
   }
 
   remove(id: string) {
-    this.employeeService.remove(id).subscribe((a) => {
-      console.log(a);
-    });
+    if (confirm('Are You Sure ?')) {
+      this.employeeService.remove(id).subscribe((a) => {
+        console.log(a);
+      });
+    }
   }
 }
